test: Add unit tests + ci to test the txfetcher (#12)
## 📝 Summary

This PR adds a unit test to check the end-to-end functionality of the
transaction fetcher. It uses an anvil node as an Eth client to subscribe
to its transactions. Adding also anvil to the CI/CD pipeline.

The setup for the test is a bit verbose now but I imagine that if this
pattern becomes useful (unit testing with a real Eth node) we will
create more concrete util abstractions.

---

## ✅ I have completed the following steps:

* [x] Run `make lint`
* [x] Run `make test`
* [x] Added tests (if applicable)