Fix: Use lock properly in filter_order_by_nonces to do it concurrently (#38)
## 📝 Summary

The current use of Mutex on `nonce_cache` in `filter_order_by_nonces` is
not correct. It causes the function to run sequentially because every
other task will wait for prev one ( both read and write ) complete in
order to do its work.
I change to use RwLock instead. Before the change it takes 18m to fetch
a block with 2k4 orders, and now it takes 18s with 400 concurrent
request.

- And lint the file

## 💡 Motivation and Context

<!--- (Optional) Why is this change required? What problem does it
solve? Remove this section if not applicable. -->

---

## ✅ I have completed the following steps:

* [x] Run `make lint`
* [x] Run `make test`
* [ ] Added tests (if applicable)