feat: Add abstraction for the beacon api client (#34)
This PR adds a small abstraction over the beacon client api and
implements the methods required by the rbuilder.

There are a couple of unit tests that require a running beacon node.
They are ignore right now but we will enable them in a followup PR.

<!--- (Optional) Why is this change required? What problem does it
solve? Remove this section if not applicable. -->

---

* [x] Run `make lint`
* [x] Run `make test`
* [x] Added tests (if applicable)