Add playground integration (#69)
## 📝 Summary

This PR implements and integration test using the
[builder-playground](https://github.com/flashbots/builder-playground):

- It creates a playground.rs utility library to interact with the
playground. **In this first iteration, most of the variables (e.g.
private keys, urls, etc...) are hardcoded**.
- It creates a simple integration test that ensures that a transaction
sent to the EL gets mined by the rbuilder and included on the block
through mev-boost-relay.
- It runs the integration test in CI after the unit tests have passed.

The integration test works directly with the `rbuilder` compiled binary.

## 💡 Motivation and Context

<!--- (Optional) Why is this change required? What problem does it
solve? Remove this section if not applicable. -->

---

## ✅ I have completed the following steps:

* [ ] Run `make lint`
* [ ] Run `make test`
* [ ] Added tests (if applicable)