Wait for rbuilder running before running integration test (#160)
## 📝 Summary

The new integration test introduced in #143 requires the rbuilder to be
running to work. However, the playground setup was not waiting for it.
This PR fixes that. It adds a loop after the rbuilder creation that
checks in the output log for a specific log "RPC server job: started"
that signals that the builder is running and the test can start.

## 💡 Motivation and Context

<!--- (Optional) Why is this change required? What problem does it
solve? Remove this section if not applicable. -->

---

## ✅ I have completed the following steps:

* [x] Run `make lint`
* [x] Run `make test`
* [x] Added tests (if applicable)