fix provider reopener, rework nonce cache
The provider factory reopener's concept of "last_consistent_block" was incorrect, as database consistency was compromised for previously checked blocks. This caused the rbuilder to miss some slots due to its inability to reopen the database during slot execution.

The nonce cache has been refactored to use database transactions instead of database handles. This change simplifies the code and reduces the need for database consistency checks.