Introduce ThreadBlockBuildingContext, removing old CachedReads.
The idea behind ThreadBlockBuildingContext is to have a place that
is owned by a particular thread (for example, max profit ordering
builder) that can be used as a scratchpad for caching without doing
global caching shared between all threads.

Currently we have only one place for caches BlockBuildingContext. It
is global and shared between all builders, top of block simulation,
and finalization thread. This approach solves many problems with caching. For
example, we use it for eth-sparse-mpt root hash caching.

But sometimes we would like to have a cache that is local to the current
thread to avoid the overhead of mutex and multiple threads.

I've seen the need for caches like this multiple times.
This commit adds support for local caches like this.

The first thing that is moved to this new setup of 2 caches is cached reads. We had a
lot of trouble with passing CachedReads struct around and it
introduced itself into traits where it does not belong such as
BlockBuildingHelper or backtesting code. Here we move CachedReads to this uniform
setup and it simplifies cached reads state handling.