Fixed finalize/roothash metrics (#812)
## 📝 Summary

Metrics were wrong since finalize_adjust_time and finalize_time were
always updated no matter if it was the first finalize call or an update.
Not we differentiate the cases and combined the metrics in a single one
with a label "mode".
Did the same differentiation for root hash times (label "mode") so we
can tell the difference between the first root hash and an adjustment.

## ✅ I have completed the following steps:

* [X] Run `make lint`
* [X] Run `make test`
* [ ] Added tests (if applicable)